refactor "respect one-time Cc/Bcc"
using radio is a bad idea, which is quite confusing and doesn't guarantee to
be vertical aligned either.

note that it's convenient to automatically update preview scrips when user
selects autocomplete users, so we need to trigger the change on select.

the css tweak is to make sure simplified recipients are vertical aligned,
the normal preview scrips part doesn't have this issue.