Pass the just-submitted squelching information to the first AJAX request
If "Add more files" is pressed, the set of selected squelched recipients
is submitted; on the sebsequent page display, however, that information
does not appear in the initial form that is serialized and passed to the
helpers; as such, the state of the checkboxes is lost.

Preserve that information by ensuring that the squelching information is
present on the page during the first page load, before the AJAX
requests, by placing it within the area to be replaced.  This ensures
that the form elements will not be duplicated once the initial AJAX
request returns, but provides the query parameters for the initial call
to jQuery('form[name=TicketUpdate]').serialize().