Switch to making POST requests, not GET requests
If passed a string, the .load() function makes a GET request; if passed
an object, it makes a POST request.  Since GET requests are limited in
query-string length by the webserver, use a POST rquest to ensure that
if the helper request is not dropped if the textarea contains a large
amount of text.